Add website and prisma starter functionality
Created base-user.js file to instantiate a user to
the base db. Defined client file to be an instance
of prisma client and exported that instance to the
rest of the project. Started db/index.js file to
add more cli tracking in the future. Added basic
query to queries.js file to get all users, and for
later use in the project. Updated basic schema.