Add signup functionality and bcrypt
Created routes for signup get and post. Added
controllers for signup that allow the user to
create a profile. Added validation to the signup
page. Added bcrypt to the password creation
process. Updated the passport controller to check
using bcrypt to compare instead. Updated queries
so that inserting a new enduser would have all the
correct data. Updated the schema so that the db
would reflect the correct columns for the new user
data inserted. Created signup view for the user
to insert data.