Update schema and add create folder functionality
Updated the schema to have a folders table and a
files table with a relation between them. Added
folder creation routes and controllers so that the
user can create folders in the db. Added queries
that allow for checking the db for repeat folders
and assist the user in creation of the folders.
Added basic styling. Added create-folder view to
facilitate the creation of folders for the user.