Add file upload/download and logout
Added functionality to the project so that the
uploading of files will change based on what type
of file is being uploaded. Added new download link
to the open-folder view so that the user can
download the file they want. Added logout
functionality so the user can end their session.